Allow passing abortsignal or opts object
