+++ src/indexFile.ts
@@ -3,9 +3,7 @@ import QuickLRU from 'quick-lru'
 import { GenericFilehandle } from 'generic-filehandle'
 import VirtualOffset from './virtualOffset'
 import Chunk from './chunk'
-export interface BaseOpts {
-  signal?: AbortSignal
-}
+
 export default abstract class IndexFile {
   public filehandle: GenericFilehandle
   public renameRefSeq: Function
